Create MainActivity.java
Purpose: This is the main "Activity" for the application, it will be the first activity to be displayed when the app is dispatched. This class, also launches another instance of an "Activity" when the "Full Results" button is clicked. Because "Networking" doesn't happen immediately, a web server can take a few seconds to respond, nonetheless a "File" can take a longer time. Android disallows networking to be done on the main thread. To fix this issue, you can use the "AsyncTask" class that allows a background thread to run when you have the main thread running. Using the "doInBackground" method will execute your commands.